+++ MoreFertilizers/MoreFertilizers/ModEntry.cs
@@ -315,7 +315,8 @@ internal sealed class ModEntry : Mod
 
     private void OnButtonPressed(object? sender, ButtonPressedEventArgs e)
     {
-        if (MenuingExtensions.IsNormalGameplay())
+        if ((e.Button.IsUseToolButton() || e.Button.IsActionButton())
+            && MenuingExtensions.IsNormalGameplay())
         {
             SpecialFertilizerApplication.ApplyFertilizer(e, this.Helper.Input);
         }