+++ TapGiantCrops/docs/Changelog.md
@@ -1,5 +1,9 @@
 ﻿Changelog
 ===============
+TOOD:
+Automate compat
+Make giant flowers produce the relevant honey?
+Add a data asset for other giant crops.
 
 ### Version 0.1.1
 * Fixes integration with More Giant Crops and JA Giant Crops.